refactor: 리팩토링
- 패키지 구조 변경
- 값 객체인 Regex에 equals(), hashCode() 오버라이드
- 사용하지 않는 메서드 삭제
- Delimiter는 불변 객체이므로 record로 변경
