+++ src/main/java/calculator/util/converter/IntegerConvertible.java
@@ -1,4 +1,4 @@
-package calculator.converter;
+package calculator.util.converter;
 
 import java.util.List;
