+++ src/main/java/calculator/util/regex/RegexGenerator.java
@@ -1,16 +1,16 @@
-package calculator.regex;
+package calculator.util.regex;
 
-import static delimiter.DelimiterPattern.CUSTOM_DELIMITER;
+import static calculator.domain.delimiter.DelimiterPattern.CUSTOM_DELIMITER;
 
-import delimiter.Delimiter;
-import delimiter.Delimiters;
+import calculator.domain.delimiter.Delimiter;
+import calculator.domain.delimiter.Delimiters;
 
 public class RegexGenerator {
 
     public Regex makeDelimiterAreaRegex(final Delimiters delimiters) {
         Regex regex = new Regex(CUSTOM_DELIMITER.getRegex());
         for (Delimiter delimiter : delimiters.getDelimiters()) {
-            regex.add(delimiter.getDelimiter());
+            regex.add(delimiter.delimiter());
         }
 
         return regex;
